# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + Gin + MySQL + JWT 的在线单词本项目",
        "title": "单词本 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "你的名字",
            "url": "http://www.example.com",
            "email": "example@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "登录获取 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "根据单词查询释义、音标、例句",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "词典模块"
                ],
                "summary": "查询单词",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的单词",
                        "name": "word",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "meanings": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "antonyms": {
                                                                        "description": "词义的反义词（可选）",
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "definitions": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object",
                                                                            "properties": {
                                                                                "antonyms": {
                                                                                    "description": "反义词（可选）",
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "type": "string"
                                                                                    }
                                                                                },
                                                                                "definition": {
                                                                                    "description": "定义文本",
                                                                                    "type": "string"
                                                                                },
                                                                                "example": {
                                                                                    "description": "例句（可选）",
                                                                                    "type": "string"
                                                                                },
                                                                                "synonyms": {
                                                                                    "description": "同义词（可选）",
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "type": "string"
                                                                                    }
                                                                                }
                                                                            }
                                                                        }
                                                                    },
                                                                    "partOfSpeech": {
                                                                        "description": "词性",
                                                                        "type": "string"
                                                                    },
                                                                    "synonyms": {
                                                                        "description": "词义的同义词（可选）",
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "string"
                                                                        }
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "phonetics": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "audio": {
                                                                        "description": "发音音频URL",
                                                                        "type": "string"
                                                                    },
                                                                    "sourceUrl": {
                                                                        "description": "来源URL",
                                                                        "type": "string"
                                                                    },
                                                                    "text": {
                                                                        "description": "音标文本",
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        },
                                                        "word": {
                                                            "description": "单词",
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "单词不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/words": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生词本模块"
                ],
                "summary": "获取生词列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Word"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生词本模块"
                ],
                "summary": "添加单词到生词本",
                "parameters": [
                    {
                        "description": "单词信息",
                        "name": "word",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddWordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/words/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生词本模块"
                ],
                "summary": "更新单词掌握度",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "单词ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新的掌握度",
                        "name": "mastery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateMasteryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生词本模块"
                ],
                "summary": "删除生词（软删除）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "单词ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.AddWordRequest": {
            "type": "object",
            "required": [
                "word"
            ],
            "properties": {
                "definition": {
                    "type": "string"
                },
                "example": {
                    "type": "string"
                },
                "phonetic": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "handler.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateMasteryRequest": {
            "type": "object",
            "required": [
                "mastery"
            ],
            "properties": {
                "mastery": {
                    "enum": [
                        0,
                        1,
                        2
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MasteryLevel"
                        }
                    ]
                }
            }
        },
        "model.MasteryLevel": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "Fuzzy": "1: 模糊",
                "Mastered": "2: 掌握",
                "Unknowned": "0: 未知"
            },
            "x-enum-descriptions": [
                "0: 未知",
                "1: 模糊",
                "2: 掌握"
            ],
            "x-enum-varnames": [
                "Unknowned",
                "Fuzzy",
                "Mastered"
            ]
        },
        "model.Word": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "definition": {
                    "description": "释义",
                    "type": "string"
                },
                "example": {
                    "description": "例句",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_reviewed_at": {
                    "description": "上次复习时间",
                    "type": "integer"
                },
                "mastery": {
                    "description": "掌握程度",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MasteryLevel"
                        }
                    ]
                },
                "phonetic": {
                    "description": "音标",
                    "type": "string"
                },
                "review_count": {
                    "description": "复习次数",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "外键关联用户",
                    "type": "integer"
                },
                "word": {
                    "description": "不能设置为unique，因为一个单词可能是多个用户的生词",
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}